+++ events_processor/database/billable_metrics.go
@@ -1,14 +1,27 @@
 package database
 
 import (
-	"github.com/getlago/lago/events-processor/models"
+	"time"
+
 	"github.com/getlago/lago/events-processor/utils"
+	"gorm.io/gorm"
 )
 
-func (db *DB) FetchBillableMetric(organizationID string, code string) utils.Result[*models.BillableMetric] {
+type BillableMetric struct {
+	ID             string         `gorm:"primaryKey;->"`
+	OrganizationID string         `gorm:"->"`
+	Code           string         `gorm:"->"`
+	FieldName      string         `gorm:"->"`
+	Expression     string         `gorm:"->"`
+	CreatedAt      time.Time      `gorm:"->"`
+	UpdatedAt      time.Time      `gorm:"->"`
+	DeletedAt      gorm.DeletedAt `gorm:"index;->"`
+}
+
+func (db *DB) FetchBillableMetric(organizationID string, code string) utils.Result[*BillableMetric] {
 	// TODO: take deleted records into account
 
-	var bm *models.BillableMetric
+	var bm *BillableMetric
 	result := db.connection.First(bm, "organization_id = ? AND code = ?", organizationID, code)
 	if result.Error != nil {
 		return failedBillabmeMetricResult(result.Error)
@@ -17,6 +30,6 @@ func (db *DB) FetchBillableMetric(organizationID string, code string) utils.Resu
 	return utils.SuccessResult(bm)
 }
 
-func failedBillabmeMetricResult(err error) utils.Result[*models.BillableMetric] {
-	return utils.FailedResult[*models.BillableMetric](err)
+func failedBillabmeMetricResult(err error) utils.Result[*BillableMetric] {
+	return utils.FailedResult[*BillableMetric](err)
 }
