+++ events_processor/database/charges.go
@@ -1,15 +1,27 @@
 package database
 
 import (
-	"github.com/getlago/lago/events-processor/models"
+	"time"
+
 	"github.com/getlago/lago/events-processor/utils"
+	"gorm.io/gorm"
 )
+
+type Charge struct {
+	ID               string         `gorm:"primaryKey;->"`
+	BillableMetricID string         `gorm:"->"`
+	PlanID           string         `gorm:"->"`
+	PayInAdvance     bool           `gorm:"->"`
+	CreatedAt        time.Time      `gorm:"->"`
+	UpdatedAt        time.Time      `gorm:"->"`
+	DeletedAt        gorm.DeletedAt `gorm:"index;->"`
+}
 
 func (db *DB) AnyInAdvanceCharge(planID string, billableMetricID string) utils.Result[bool] {
 	// TODO: take deleted records into account
 	var count int64
 
-	result := db.connection.Model(&models.Charge{}).
+	result := db.connection.Model(&Charge{}).
 		Where("plan_id = ? AND billable_metric_id = ?", planID, billableMetricID).
 		Count(&count)
 	if result.Error != nil {
