--- events_processor/models/billable_metric.go
@@ -1,18 +1,0 @@
-package models
-
-import (
-	"time"
-
-	"gorm.io/gorm"
-)
-
-type BillableMetric struct {
-	ID             string         `gorm:"primaryKey;->"`
-	OrganizationID string         `gorm:"->"`
-	Code           string         `gorm:"->"`
-	FieldName      string         `gorm:"->"`
-	Expression     string         `gorm:"->"`
-	CreatedAt      time.Time      `gorm:"->"`
-	UpdatedAt      time.Time      `gorm:"->"`
-	DeletedAt      gorm.DeletedAt `gorm:"index;->"`
-}
