+++ events_processor/processor/models/event.go
